remove duplicate code in `Variant`
